% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyjs-package.R, R/tidyjs.R
\docType{package}
\name{tidyjs-package}
\alias{tidyjs}
\alias{tidyjs-package}
\title{tidyjs: Tidy Data with JavaScript and 'tidy.js'}
\description{
Provides the 'tidy.js' JavaScript package for use
    in HTML 'R Markdown' document a≥nd 'Shiny' applications.

Provides the 'tidy.js' JavaScript package for use
    in HTML 'R Markdown' document a≥nd 'Shiny' applications.
}
\section{Usage}{

The primary function for use in R Markdown documents or Shiny apps is
\code{\link[=use_tidyjs]{use_tidyjs()}}.
}

\references{
\url{https://pbeshai.github.io/tidy}
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Other contributors:
\itemize{
  \item Peter Beshai (Author of tidy.js) [copyright holder]
  \item Mike Bostock (Author of d3-array.js) [copyright holder]
}


\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Other contributors:
\itemize{
  \item Peter Beshai (Author of tidy.js) [copyright holder]
  \item Mike Bostock (Author of d3-array.js) [copyright holder]
}

}
\keyword{internal}
