% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyjs.R
\name{use_tidyjs}
\alias{use_tidyjs}
\alias{html_dependency_tidyjs}
\alias{html_dependency_d3_array}
\title{Use tidyjs}
\usage{
use_tidyjs(minified = TRUE, remote = FALSE)

html_dependency_tidyjs(minified = TRUE, remote = FALSE)

html_dependency_d3_array(minified = TRUE, remote = FALSE)
}
\arguments{
\item{minified}{Use the minified JavaScript files}

\item{remote}{When \code{TRUE}, only the versions hosted on
\url{https://www.unpkg.com} will be used.}
}
\description{
Adds the HTML dependencies required for
\href{https://pbeshai.github.io/tidy}{tidy.js}.
}
\details{
To use \code{tidy.js} in your R Markdown documents, call
\code{\link[=use_tidyjs]{use_tidyjs()}}. You can then access tidy functions from
the \code{Tidy} object in your Shiny App or in your R Markdown document
inside a JavaScript (\code{js}) chunk:\if{html}{\out{<div class="markdown">}}\preformatted{```\{r echo=FALSE\}
tidyjs::use_tidyjs()
```

```\{js\}
const \{ tidy, mutate, arrange, desc \} = Tidy;

const data = [
  \{ a: 1, b: 10 \}, 
  \{ a: 3, b: 12 \}, 
  \{ a: 2, b: 10 \}
]

const results = tidy(
  data, 
  mutate(\{ ab: d => d.a * d.b \}),
  arrange(desc('ab'))
)
```
}\if{html}{\out{</div>}}
}
\section{Functions}{
\itemize{
\item \code{use_tidyjs}: Add the \code{tidy.js} and dependencies to your document
or app

\item \code{html_dependency_tidyjs}: The \code{tidy.js} dependency

\item \code{html_dependency_d3_array}: The \code{d3-array} dependency
}}

